Incorporated localCredentials into user.
