--- src/main/java/entity/LocalCredentialsInterface.java
@@ -1,6 +1,0 @@
-package entity;
-
-public interface LocalCredentialsInterface {
-    String getUsername();
-    byte[] getHashedPassword();
-}
